Fixed the logic in deleteByNameAndType() and deletePool() and added the test case deleteAlsoRemovesPool()
* Fixed the logic in deleteByNameAndType() in UltraDNSRestResourceRecordSetApi
* Fixed the logic in deletePool() in UltraDNSRestRoundRobinPoolApi
* Added the test case deleteAlsoRemovesPool() in UltraDNSRestResourceRecordSetApiMockTest

+++ ultradns-rest/src/main/java/denominator/ultradns/UltraDNSRestResourceRecordSetApi.java
@@ -152,20 +152,12 @@ final class UltraDNSRestResourceRecordSetApi implements denominator.ResourceReco
 
   @Override
   public void deleteByNameAndType(String name, String type) {
-    try {
-      api.deleteResourceRecordByNameType(zoneName, lookup(type), name);
-    } catch (UltraDNSRestException e) {
-      if (e.code() != UltraDNSRestException.RESOURCE_RECORD_NOT_FOUND) {
-        throw e;
-      }
-    }
-    if (roundRobinPoolApi.isPoolType(type)) {
-      roundRobinPoolApi.deletePool(name, type);
+    for (Record record : recordsByNameAndType(name, type)) {
+      remove(name, type, record);
     }
   }
 
   private void remove(String name, String type, Record record) {
-    int indexToDelete = -1;
     String rData = "";
     int intType = lookup(type);
 
@@ -173,30 +165,35 @@ final class UltraDNSRestResourceRecordSetApi implements denominator.ResourceReco
       rData = StringUtils.join(record.getRdata(), " ");
     }
 
+    List<RRSet> rrSets = null;
     try {
-      List<RRSet> rrSets = api.getResourceRecordsOfDNameByType(zoneName, name, intType).getRrSets();
-      if (rrSets != null && !rrSets.isEmpty()) {
-        RRSet rrSet = rrSets.get(0);
-        if (rrSet != null & rrSet.getRdata() != null) {
-          indexToDelete = rrSet.getRdata().indexOf(rData);
-        }
-      }
+      rrSets = api.getResourceRecordsOfDNameByType(zoneName, name, intType).getRrSets();
     } catch (UltraDNSRestException e) {
       if (e.code() != UltraDNSRestException.DATA_NOT_FOUND) {
         throw e;
       }
     }
 
-    if (indexToDelete >= 0 ) {
-      try {
-        api.deleteResourceRecord(zoneName, intType, name, indexToDelete);
-      } catch (UltraDNSRestException e) {
-        if (e.code() != UltraDNSRestException.PATH_NOT_FOUND_TO_PATCH) {
-          throw e;
+    RRSet rrSet = null;
+    if (rrSets != null && !rrSets.isEmpty()) {
+      rrSet = rrSets.get(0);
+      if (rrSet != null & rrSet.getRdata() != null) {
+        int indexToDelete = rrSet.getRdata().indexOf(rData);
+        if (indexToDelete >= 0 ) {
+          try {
+            api.deleteResourceRecord(zoneName, intType, name, indexToDelete);
+          } catch (UltraDNSRestException e) {
+            if (e.code() != UltraDNSRestException.PATH_NOT_FOUND_TO_PATCH) {
+              throw e;
+            }
+          }
         }
       }
     }
-    if (roundRobinPoolApi.isPoolType(type)) {
+
+    // If the last record in the pool is deleted successfully or if there are
+    // no records in the pool and if it is a pool then delete the pool itself.
+    if (rrSet != null && rrSet.getRdata().size() <= 1 && roundRobinPoolApi.isPoolType(type)) {
       roundRobinPoolApi.deletePool(name, type);
     }
   }
